+++ ITI.Human.Data/Res/Model/ITIH.OrderFinalDueTable.Install.1.0.0.sql
@@ -1,8 +1,8 @@
 ﻿create table ITIH.tOrderFinalDue (
 	OrderFinalDueId int not null identity(0, 1),
 	OrderId int not null,
-	Total float not null,
-	Paid float not null default 0
+	Total int not null,
+	Paid int not null default 0
 
 	constraint PK_ITIH_tOrderFinalDue_OrderFinalDueId primary key (OrderFinalDueId),
 	constraint FK_ITIH_tOrderFinalDue_OrderId foreign key (OrderId) references ITIH.tOrder (OrderId)
