+++ ITI.Human.Data/Res/Model/ITIH.StorageLinkedProductTable.Install.1.0.0.sql
@@ -2,7 +2,7 @@
 	StorageLinkedProductId int not null identity(0, 1),
 	StorageId int not null,
 	ProductId int not null,
-	UnitPrice float not null default 0,
+	UnitPrice int not null default 0,
 	Stock int not null default 0
 
 	constraint PK_tStorageLinkedProduct_StorageLinkedProductId primary key (StorageLinkedProductId),
