+++ ITI.Human.ViewModels/Storage/LinkedProduct/CreationViewModel.cs
@@ -15,7 +15,7 @@
         /// <summary>
         /// See <see cref="BasicDataStorageLinkedProduct.UnitPrice"/>.
         /// </summary>
-        public double UnitPrice { get; set; }
+        public int UnitPrice { get; set; }
 
         /// <summary>
         /// See <see cref="BasicDataStorageLinkedProduct.Stock"/>.
